integrate and pass latest tests (#337)
* update submodules

* remove jamixir

* fix path

* fix codec test

* fix stats tests

* update jamtestvectors

* fix a few tests

* fix encoding issue

* comment out some test

* update submodule

* various fix and new logs

* fix log

* various fix

* update submodules

* pass fallback and safrole trace

* various fix and comformance

* fix decoder tests

* service accounts fix

* encode data fix

* fix sbrk and int32 usage

* fix info

* fix memory test

* fix accumulate stats

* fix state trie

* fix footprint

* fix preimage update

* fix some host calls

* update trace tests

* pass trace tests

+++ Blockchain/Sources/Blockchain/VMInvocations/Invocations/AccumulateInvocation.swift
@@ -31,7 +31,7 @@ public func accumulate(
         return .init(state: state, transfers: [], commitment: nil, gasUsed: Gas(0), provide: [])
     }
 
-    let initialIndex = try Blake2b256.hash(JamEncoder.encode(serviceIndex, state.entropy, timeslot)).data.decode(UInt32.self)
+    let initialIndex = try Blake2b256.hash(JamEncoder.encode(UInt(serviceIndex), state.entropy, UInt(timeslot))).data.decode(UInt32.self)
     let nextAccountIndex = try await AccumulateContext.check(
         i: initialIndex % serviceIndexModValue + 256,
         accounts: state.accounts.toRef()
@@ -50,7 +50,7 @@ public func accumulate(
         )
     )
     let ctx = AccumulateContext(context: contextContent, config: config, timeslot: timeslot, operands: arguments)
-    let argument = try JamEncoder.encode(timeslot, serviceIndex, arguments.count)
+    let argument = try JamEncoder.encode(UInt(timeslot), UInt(serviceIndex), UInt(arguments.count))
 
     let (exitReason, gas, output) = await invokePVM(
         config: config,
